requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
xgboost>=2.0
torch>=2.0
prophet>=1.1
